Refactor: create runtime/scheduler/ and extract scheduler types
- Move pto_scheduler.h/.cpp into runtime/scheduler/ subdirectory
- Create scheduler_types.h with CoreExecState, CoreTracker,
  SlotTransition, SchedProfilingCounters, SyncStartDrainState,
  profiling macros and scheduler constants
- Update include paths to "scheduler/pto_scheduler.h" in 3 files
- Switch CMake from GLOB to GLOB_RECURSE for source collection
  to support subdirectories without build_config.py changes
- Remove duplicate type definitions from aicpu_executor.cpp
- Add missing copyright headers to onboard CMakeLists.txt files

Pure restructuring — no behavioral change.
