+++ src/a2a3/platform/sim/aicpu/CMakeLists.txt
@@ -45,7 +45,7 @@ list(APPEND AICPU_SOURCES ${COMMON_SOURCES})
 # Then, collect sources from CUSTOM_SOURCE_DIRS (runtime sources)
 if(DEFINED CUSTOM_SOURCE_DIRS)
     foreach(SRC_DIR ${CUSTOM_SOURCE_DIRS})
-        file(GLOB DIR_SOURCES "${SRC_DIR}/*.cpp" "${SRC_DIR}/*.c")
+        file(GLOB_RECURSE DIR_SOURCES "${SRC_DIR}/*.cpp" "${SRC_DIR}/*.c")
         list(APPEND AICPU_SOURCES ${DIR_SOURCES})
     endforeach()
 else()
