+++ src/a2a3/runtime/tensormap_and_ringbuffer/runtime/pto_runtime2.h
@@ -40,7 +40,7 @@
 #include "pto_shared_memory.h"
 #include "pto_ring_buffer.h"
 #include "pto_tensormap.h"
-#include "pto_scheduler.h"
+#include "scheduler/pto_scheduler.h"
 #include "pto_orchestrator.h"
 
 // =============================================================================
